requirements.txt
aiocache==0.12.3
aiohappyeyeballs==2.5.0
aiohttp==3.10.11
aiolimiter==1.2.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.11
fastapi-offline==1.7.3
Flask==2.3.3
frozenlist==1.5.0
gevent==24.11.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.1.0
numpy==1.26.0
pandas==2.0.3
pandas_ta==0.3.14b0
passlib==1.7.4
propcache==0.3.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.1
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.44.0
ta==0.10.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.33.0
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.3
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
